+++ monolith_service/routers/produce.py
@@ -4,11 +4,9 @@ from urllib import response
 from fastapi import APIRouter, Depends, Response
 from fastapi.encoders import jsonable_encoder
 from queries.produce import Error, Produce_update_available, ProduceQueries, Produce_get, Produce_create, ProduceDataClass, ProduceRequest
-
+from authenticator import authenticator
 
 router = APIRouter()
-
-
 
 
 @router.get("/api/produce/", response_model = list[Produce_get]) 
@@ -20,9 +18,11 @@ def get_all_produce(
 @router.post("/api/produce/", response_model = Produce_get)
 def create_produce(
     produce: Produce_create,
-    queries: ProduceQueries = Depends()
+    queries: ProduceQueries = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data)
 ):
-    return queries.create_produce(produce)
+    if "admin" in account_data.get("roles"):
+        return queries.create_produce(produce)
 
 @router.get("/api/produce/{produce_id}", response_model = Optional[Produce_get])
 def get_single_produce_item(
@@ -40,15 +40,19 @@ def update_produce(
     produce_id: int,
     produce: Produce_create,
     queries: ProduceQueries = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data)
 ) -> Union[Produce_get, Error]:
-    return queries.update_produce(produce_id, produce)
+    if "admin" in account_data.get("roles"):
+        return queries.update_produce(produce_id, produce)
 
 @router.delete("/api/produce/{produce_id}/delete", response_model = bool)
 def delete_produce(
     produce_id: int,
     queries: ProduceQueries = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data)
 )-> bool:
-    return queries.delete_produce(produce_id)
+    if "admin" in account_data.get("roles"):
+        return queries.delete_produce(produce_id)
 
 
 @router.patch("/api/produce/{produce_id}/patch", response_model = Produce_get)
@@ -56,12 +60,14 @@ def update_produce_available(
     produce_id: int,
     produce: Produce_update_available,
     queries: ProduceQueries = Depends(),
+    account_data: dict = Depends(authenticator.get_current_account_data)
 ) -> Produce_get:
-    # print(produce)
-    return queries.update_produce_available(produce_id, Produce_update_available(
-        available = produce.available,
-        height = produce.height, 
-        length = produce.length, 
-        width = produce.width
-        ))
+    if "admin" in account_data.get("roles"):
+        # print(produce)
+        return queries.update_produce_available(produce_id, Produce_update_available(
+            available = produce.available,
+            height = produce.height, 
+            length = produce.length, 
+            width = produce.width
+            ))
